remove Locations etc and then Shop updated
