+++ pom.xml
@@ -106,11 +106,16 @@
 				</exclusion>
 			</exclusions>
 		</dependency>
-<dependency>
-    <groupId>com.google.maps</groupId>
-    <artifactId>google-maps-services</artifactId>
-    <version>0.15.0</version>
+		<dependency>
+  <groupId>com.google.maps</groupId>
+  <artifactId>google-maps-services</artifactId>
+  <version>0.15.0</version>
 </dependency>
+<dependency>
+  <groupId>org.slf4j</groupId>
+  <artifactId>slf4j-simple</artifactId>
+</dependency>
+
 <dependency>  
     <groupId>org.apache.tiles</groupId>  
     <artifactId>tiles-servlet</artifactId>
