+++ src/main/java/com/ucs/edu/Review/config/SecurityConfig.java
@@ -43,7 +43,8 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         		"/profile","/header","/footer",
         		"/create_shop","/shop_list","/save_shop","/cate",
         		"/create_brand","/brand_list","/save_brand",
-        		"/create_product","/product_list","/save_product","/product_detail","/product/**"
+        		"/create_product","/product_list","/save_product","/product_detail",
+        		"/product/**","/productsbycategory/**"
         		,"/category_list","/save_category","/create_category"
         		,"/blog-details.htm","/faq.htm","/blog.htm",
         		"/create_location","/save_location","/location_list",
