--- src/main/java/com/ucs/edu/Review/dto/LocationDTO.java
@@ -1,46 +1,0 @@
-package com.ucs.edu.Review.dto;
-
-public class LocationDTO {
-	private Long location_id;
-	private String name;
-	private String address;
-	private Float lat;
-	private Float lng;
-	private String type;
-	public Long getLocation_id() {
-		return location_id;
-	}
-	public void setLocation_id(Long location_id) {
-		this.location_id = location_id;
-	}
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public String getAddress() {
-		return address;
-	}
-	public void setAddress(String address) {
-		this.address = address;
-	}
-	public Float getLat() {
-		return lat;
-	}
-	public void setLat(Float lat) {
-		this.lat = lat;
-	}
-	public Float getLng() {
-		return lng;
-	}
-	public void setLng(Float lng) {
-		this.lng = lng;
-	}
-	public String getType() {
-		return type;
-	}
-	public void setType(String type) {
-		this.type = type;
-	}
-}
