+++ src/main/java/com/ucs/edu/Review/dto/ShopDTO.java
@@ -1,9 +1,43 @@
 package com.ucs.edu.Review.dto;
 
+import java.util.Set;
+
+import javax.persistence.OneToMany;
+
+import com.ucs.edu.Review.model.Product;
+
 public class ShopDTO {
 	private Long shop_id;
 	private String shop_name;
-	private Long location_id;
+	private String address;
+	private Float lat;
+	private Float lng;
+	private String type;
+	
+	public String getAddress() {
+		return address;
+	}
+	public void setAddress(String address) {
+		this.address = address;
+	}
+	public Float getLat() {
+		return lat;
+	}
+	public void setLat(Float lat) {
+		this.lat = lat;
+	}
+	public Float getLng() {
+		return lng;
+	}
+	public void setLng(Float lng) {
+		this.lng = lng;
+	}
+	public String getType() {
+		return type;
+	}
+	public void setType(String type) {
+		this.type = type;
+	}
 	public Long getShop_id() {
 		return shop_id;
 	}
@@ -15,12 +49,6 @@ public class ShopDTO {
 	}
 	public void setShop_name(String shop_name) {
 		this.shop_name = shop_name;
-	}
-	public Long getLocation_id() {
-		return location_id;
-	}
-	public void setLocation_id(Long location_id) {
-		this.location_id = location_id;
 	}
 	
 }
