--- src/main/java/com/ucs/edu/Review/model/Location.java
@@ -1,82 +1,0 @@
-package com.ucs.edu.Review.model;
-
-import java.io.Serializable;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-@Entity
-public class Location implements Serializable{
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long location_id;
-	private String name;
-	private String address;
-	private Float lat;
-	private Float lng;
-	private String type;
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public String getAddress() {
-		return address;
-	}
-	public void setAddress(String address) {
-		this.address = address;
-	}
-	public Float getLat() {
-		return lat;
-	}
-	public void setLat(Float lat) {
-		this.lat = lat;
-	}
-	public Float getLng() {
-		return lng;
-	}
-	public void setLng(Float lng) {
-		this.lng = lng;
-	}
-	public String getType() {
-		return type;
-	}
-	public void setType(String type) {
-		this.type = type;
-	}
-	public Long getLocation_id() {
-		return location_id;
-	}
-	public Location() {
-		super();
-	}
-	public Location(Long location_id, String name, String address, Float lat, Float lng, String type, Shop shop) {
-		super();
-		this.location_id = location_id;
-		this.name = name;
-		this.address = address;
-		this.lat = lat;
-		this.lng = lng;
-		this.type = type;
-		this.shop = shop;
-	}
-	public void setLocation_id(Long location_id) {
-		this.location_id = location_id;
-	}
-	@OneToOne(mappedBy = "location")
-	private Shop shop;
-	public Shop getShop() {
-		return shop;
-	}
-	public void setShop(Shop shop) {
-		this.shop = shop;
-	}
-}
