+++ src/main/java/com/ucs/edu/Review/service/ShopService.java
@@ -3,11 +3,9 @@ package com.ucs.edu.Review.service;
 import java.util.List;
 
 import com.ucs.edu.Review.dto.ShopDTO;
-import com.ucs.edu.Review.model.Location;
 import com.ucs.edu.Review.model.Shop;
 
 public interface ShopService {
 	public void saveShop(ShopDTO shopDTO);
 	public List<Shop> getShopList();
-	public List<Location> getLocationList();
 }
