+++ src/main/webapp/WEB-INF/create_location.jsp
@@ -6,7 +6,7 @@
 <html>
 <head>
 <meta charset="ISO-8859-1">
-<title>Location</title>
+<title>Add Shop</title>
 </head>
 <body>
 <%@ include file="bootstrap.jsp"%>
